pyproject.toml
[project]
name = "Cyberpunk-Helper-Scripts"
version = "0.0.1"
description = "A collection of scripts to help with Cyberpunk 2077 modding."
readme = "README.md"
requires-python = "==3.11.10"
dependencies = [
    'pillow',
    'numpy',
    "pydantic>=2.11.4",
]
[project.optional-dependencies]
blender = [
    'bpy',
    'mathutils',
]
